+++ cognite/neat/_rules/importers/_rdf/_inference2rules.py
@@ -10,7 +10,7 @@ from cognite.client import data_modeling as dm
 from rdflib import RDF, RDFS, Graph, Namespace, URIRef
 from rdflib import Literal as RdfLiteral
 
-from cognite.neat._constants import NEAT
+from cognite.neat._constants import NEAT, get_default_prefixes_and_namespaces
 from cognite.neat._issues import IssueList
 from cognite.neat._issues.warnings import PropertyValueTypeUndefinedWarning
 from cognite.neat._rules.analysis import InformationAnalysis
@@ -81,7 +81,7 @@ class InferenceImporter(BaseRDFImporter):
     def from_graph_store(
         cls,
         store: NeatGraphStore,
-        data_model_id: (dm.DataModelId | tuple[str, str, str]) = DEFAULT_INFERENCE_DATA_MODEL_ID,
+        data_model_id: dm.DataModelId | tuple[str, str, str] = DEFAULT_INFERENCE_DATA_MODEL_ID,
         max_number_of_instance: int = -1,
         non_existing_node_type: UnknownEntity | AnyURI = DEFAULT_NON_EXISTING_NODE_TYPE,
         language: str = "en",
@@ -317,23 +317,6 @@ class SubclassInferenceImporter(BaseRDFImporter):
         data_types.Float.as_xml_uri_ref(): data_types.Double.as_xml_uri_ref(),
     }
 
-    def __init__(
-        self,
-        issue_list: IssueList,
-        graph: Graph,
-        rules: InformationRules,
-        non_existing_node_type: UnknownEntity | AnyURI = DEFAULT_NON_EXISTING_NODE_TYPE,
-    ) -> None:
-        super().__init__(
-            issue_list,
-            graph,
-            rules.metadata.as_data_model_id().as_tuple(),  # type: ignore[arg-type]
-            -1,
-            non_existing_node_type,
-            language="en",
-        )
-        self._rules = rules
-
     _ordered_class_query = """SELECT DISTINCT ?class (count(?s) as ?instances )
                            WHERE { ?s a ?class }
                            group by ?class order by DESC(?instances)"""
@@ -367,25 +350,57 @@ class SubclassInferenceImporter(BaseRDFImporter):
                               }}
                             }}"""
 
+    def __init__(
+        self,
+        issue_list: IssueList,
+        graph: Graph,
+        rules: InformationRules | None = None,
+        data_model_id: dm.DataModelId | tuple[str, str, str] | None = None,
+        non_existing_node_type: UnknownEntity | AnyURI = DEFAULT_NON_EXISTING_NODE_TYPE,
+    ) -> None:
+        if sum([1 for v in [rules, data_model_id] if v is not None]) != 1:
+            raise ValueError("Exactly one of rules or data_model_id must be provided.")
+        if data_model_id is not None:
+            identifier = data_model_id
+        elif rules is not None:
+            identifier = rules.metadata.as_data_model_id().as_tuple()  # type: ignore[assignment]
+        else:
+            raise ValueError("Exactly one of rules or data_model_id must be provided.")
+        super().__init__(issue_list, graph, identifier, -1, non_existing_node_type, language="en")
+        self._rules = rules
+
     def _to_rules_components(
         self,
     ) -> dict:
+        if self._rules:
+            prefixes = self._rules.prefixes.copy()
+        else:
+            prefixes = get_default_prefixes_and_namespaces()
+
         parent_by_child = self._read_parent_by_child_from_graph()
         read_properties = self._read_class_properties_from_graph(parent_by_child)
-        classes, properties = self._create_classes_properties(read_properties)
+        classes, properties = self._create_classes_properties(read_properties, prefixes)
 
+        if self._rules:
+            metadata = self._rules.metadata.model_dump()
+            default_space = self._rules.metadata.prefix
+        else:
+            metadata = self._default_metadata()
+            default_space = metadata["space"]
         return {
-            "metadata": self._rules.metadata.model_dump(),
-            "classes": [cls.dump(self._rules.metadata.prefix) for cls in classes],
-            "properties": [prop.dump(self._rules.metadata.prefix) for prop in properties],
-            "prefixes": self._rules.prefixes,
+            "metadata": metadata,
+            "classes": [cls.dump(default_space) for cls in classes],
+            "properties": [prop.dump(default_space) for prop in properties],
+            "prefixes": prefixes,
         }
 
     def _create_classes_properties(
-        self, read_properties: list[_ReadProperties]
+        self, read_properties: list[_ReadProperties], prefixes: dict[str, Namespace]
     ) -> tuple[list[InformationInputClass], list[InformationInputProperty]]:
-        existing_classes = {class_.class_.suffix: class_ for class_ in self._rules.classes}
-        prefixes = self._rules.prefixes.copy()
+        if self._rules:
+            existing_classes = {class_.class_.suffix: class_ for class_ in self._rules.classes}
+        else:
+            existing_classes = {}
         classes: list[InformationInputClass] = []
         properties: list[InformationInputProperty] = []
         # Help for IDE
@@ -399,20 +414,22 @@ class SubclassInferenceImporter(BaseRDFImporter):
                 parent_class_properties_iterable
             )
 
-            shared_property_uris = set.intersection(
-                *[
-                    set(properties_by_property.keys())
-                    for properties_by_property in properties_by_class_by_property.values()
-                ]
-            )
+            parent_suffix: str | None = None
             if parent_uri != NEAT.EmptyType:
+                shared_property_uris = set.intersection(
+                    *[
+                        set(properties_by_property.keys())
+                        for properties_by_property in properties_by_class_by_property.values()
+                    ]
+                )
                 parent_suffix = remove_namespace_from_uri(parent_uri)
                 self._add_uri_namespace_to_prefixes(parent_uri, prefixes)
                 if parent_suffix not in existing_classes:
                     classes.append(InformationInputClass(class_=parent_suffix))
                 else:
                     classes.append(InformationInputClass.load(existing_classes[parent_suffix].model_dump()))
-
+            else:
+                shared_property_uris = set()
             shared_properties: dict[URIRef, list[_ReadProperties]] = defaultdict(list)
             for type_uri, properties_by_property_uri in properties_by_class_by_property.items():
                 class_suffix = remove_namespace_from_uri(type_uri)
@@ -422,7 +439,7 @@ class SubclassInferenceImporter(BaseRDFImporter):
                     classes.append(
                         InformationInputClass(
                             class_=class_suffix,
-                            implements=parent_suffix if parent_uri != NEAT.UnknownType else None,
+                            implements=parent_suffix,
                         )
                     )
                 else:
@@ -434,12 +451,14 @@ class SubclassInferenceImporter(BaseRDFImporter):
                     properties.append(
                         self._create_property(read_properties, class_suffix, type_uri, property_uri, prefixes)
                     )
-            for property_uri, read_properties in shared_properties.items():
-                properties.append(
-                    self._create_property(
-                        read_properties, parent_suffix, read_properties[0].type_uri, property_uri, prefixes
+
+            if parent_suffix:
+                for property_uri, read_properties in shared_properties.items():
+                    properties.append(
+                        self._create_property(
+                            read_properties, parent_suffix, read_properties[0].type_uri, property_uri, prefixes
+                        )
                     )
-                )
         return classes, properties
 
     @staticmethod
@@ -461,15 +480,17 @@ class SubclassInferenceImporter(BaseRDFImporter):
         for result_row in self.graph.query(self._ordered_class_query):
             type_uri, instance_count_literal = cast(tuple[URIRef, RdfLiteral], result_row)
             count_by_type[type_uri] = instance_count_literal.toPython()
-        del result_row
-        analysis = InformationAnalysis(self._rules)
-        existing_class_properties = {
-            (class_entity.suffix, prop.property_)
-            for class_entity, properties in analysis.classes_with_properties(
-                consider_inheritance=True, allow_different_namespace=True
-            ).items()
-            for prop in properties
-        }
+        if self._rules:
+            analysis = InformationAnalysis(self._rules)
+            existing_class_properties = {
+                (class_entity.suffix, prop.property_)
+                for class_entity, properties in analysis.classes_with_properties(
+                    consider_inheritance=True, allow_different_namespace=True
+                ).items()
+                for prop in properties
+            }
+        else:
+            existing_class_properties = set()
         properties_by_class_by_subclass: list[_ReadProperties] = []
         for type_uri, instance_count in count_by_type.items():
             property_query = self._properties_query.format(type=type_uri, unknown_type=NEAT.UnknownType)
@@ -487,7 +508,6 @@ class SubclassInferenceImporter(BaseRDFImporter):
                 if occurrence_row:
                     max_occurrence_literal, *__ = cast(tuple[RdfLiteral, Any], occurrence_row)
                     max_occurrence = int(max_occurrence_literal.toPython())
-                del occurrence_row
                 properties_by_class_by_subclass.append(
                     _ReadProperties(
                         type_uri=type_uri,
@@ -498,8 +518,6 @@ class SubclassInferenceImporter(BaseRDFImporter):
                         instance_count=instance_count,
                     )
                 )
-
-            del result_row
         return properties_by_class_by_subclass
 
     def _read_parent_by_child_from_graph(self) -> dict[URIRef, URIRef]:
@@ -507,7 +525,6 @@ class SubclassInferenceImporter(BaseRDFImporter):
         for result_row in self.graph.query(self._type_parent_query):
             parent_uri, child_uri = cast(tuple[URIRef, URIRef], result_row)
             parent_by_child[child_uri] = parent_uri
-        del result_row
         return parent_by_child
 
     def _create_property(
@@ -534,7 +551,7 @@ class SubclassInferenceImporter(BaseRDFImporter):
     def _get_value_type(
         self, read_properties: list[_ReadProperties], prefixes: dict[str, Namespace]
     ) -> str | UnknownEntity:
-        value_types = {prop.value_type for prop in read_properties}
+        value_types = {self.overwrite_data_types.get(prop.value_type, prop.value_type) for prop in read_properties}
         if len(value_types) == 1:
             uri_ref = value_types.pop()
             if uri_ref == NEAT.UnknownType:
@@ -546,3 +563,16 @@ class SubclassInferenceImporter(BaseRDFImporter):
         for uri_ref in value_types:
             self._add_uri_namespace_to_prefixes(uri_ref, prefixes)
         return " | ".join(remove_namespace_from_uri(uri_ref) for uri_ref in value_types)
+
+    def _default_metadata(self) -> dict[str, Any]:
+        now = datetime.now(timezone.utc)
+        return InformationMetadata(
+            space=self.data_model_id.space,
+            external_id=self.data_model_id.external_id,
+            version=cast(str, self.data_model_id.version),
+            name="Inferred Model",
+            creator=["NEAT"],
+            created=now,
+            updated=now,
+            description="Inferred model from knowledge graph",
+        ).model_dump()